Refactored bulk of pipeline code into a domain class
